+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All notable changes to this project will be documented in this file. This change
 - Added new arity to `uuid` to allow you to pass a string. The string is then SHA-1 hashed to a UUID. Only valid in RethinkDB >= 2.2. [#109](https://github.com/apa512/clj-rethinkdb/issues/109)
 - Added new RethinkDB term `values`. This is analogous to Clojure's `vals` function. Only valid in RethinkDB >= 2.2. [#110](https://github.com/apa512/clj-rethinkdb/issues/110)
 - Added new RethinkDB query term `server`. This is used to retrieve information about the server you are connected to. [#111](https://github.com/apa512/clj-rethinkdb/issues/111)
+- Add new RethinkDB query term `reconfigure`. [#78](https://github.com/apa512/clj-rethinkdb/issues/78)
+- Add new RethinkDB query term `wait`. [#78](https://github.com/apa512/clj-rethinkdb/issues/78)
 
 ### Removed
 - Support for RethinkDB < 2.0. [Changes](https://github.com/rethinkdb/rethinkdb/issues/5124) in RethinkDB 2.2 mean we can't (easily) support changefeeds in versions older than 2.0. The rest of the driver commands should work, but explicit support is being removed.